Remove shared from the tier order
The size of the tier order array is used to determine the size of subnets,
so adding shared to the known tiers resulted in unnecessarily smaller
subnets.

The only template type to use this teir is product, which explicitly
looks for the shared tier so order is not relevant in this case.

However the tier entry in the master data is still needed so it is a
known tier.