Linked SSL certificate is now optional
If an ALB has an associated link to an SSL certificate, only use
this if the certificate has been defined.

This allows the certificate entry to be in the solution, but the
certificate to be defined in a subset of the environments.

As an example, the segment wildcard certificate can be used in
development environments, but a dedicated certificate can be used in
production.