Add more complete Swagger/API Gateway Integration
Add script to generate an extended swagger spec.
Permit control of integration via regular expression matching of paths
and verbs. The order the match patterns are listed is used as the
precedence order.

While this script is intended to be run in the context of an account
tree, it is likely the source swagger spec will exist in a different
repo.