Add initial support for typed components
Component type can now be provided as part of the attribute name in the
solution. This permits related components to have the same id but
different types.

For backwards compatability, the type can either be provided explicitly
via the component attribute name or implicitly via inclusion of the type
specific attribute  for the component. If provided explicitly, then any
impicit type must agree with the explicit type.