Change signature of getKey
Previously getKey didn't explicitly return a value if a key wasn't
found. It was assumed null would be returned but this isn't guaranteed.
getKey now exsplicitly returns an empty string if no key is found and
calling code should check for this if determining whether a value
exists. The most pleasing way to do this is using the ?has_content
operator.