Add lambda function dependency on role
this would seem redundant as the role is explicitly mentioned in the lambda template.
However, in attempting to upgrade the role to permit vpc access, cloud formation
is not waiting for the role to be updated before attempting to update the function.
This is an attempt to force it to do so - it doesn't hurt anyway.