Introduce "resource" abstraction
Resource layer is intended to allow replacement of generation
artefact based on framework, initially to allow prototyping of
Terraform. The heavier use of freemarker data structures is also
part of refactoring to permit generation of "whole of solution"
templates e.g. for IAM, Log Groups, and the generation of a "state"
document that reflects the config and current resource state of an
environment.

The bash scripts have also been refactored to move much of the repeated
logic into functions contained in common.sh. In addition, simple if
statements have been replaced with use of the &&/|| resulting in more
compact code. It is hopefully more readable too, especially with the
heavier use of functions.

Segment bucket names have been switched to use the product-segment style
of naming rather than incorporating the segment domain, as this has
proved problematic when changes of doamin name are required for vpcs.
The segment buckets include the vpc id in corder to ensure uniqueness
within AWS, and the newer naming naturally groups product related
buckets together.

There's more work to do on the refactoring front, but there is
sufficient change here to want to get it into the master branch.

Changes have been made in a backwards compatible way, particularly
around the preversation of deprecated id formats if the associated
resources have already been created. New environments should get all the
latest naming conventions.