Code array for weird bash 4.3 bug
If the array are declared with just -a, then they are local to the
function they are in. In this case, an attempt tocreate a nameref to
these arrays gets an error.

If however the same thing is done outside a function, it works
correctly.

The workaround is to include a "g" switch when declaring the array to
force it to be global.