Support peculiarities of MINGW64
Paths for mingw64 need to explicitly mention of the drive to ensure they
are absolute, so that the file:// operator used by the AWS CLI works ok.

By default, c:\tmp will be used as the temporary drive on windows, but
it can be explicitly set by exporting GENERATION_TMPDIR before running
gen3.