Move api gateway lambda permissions to lambda component (#119)
* Move generation of lambda permissions

Use presence of link on lambda to an API gateway to trigger the creation of
the resource permission on the lambda fn to allow this to happen.

this will allow lambda functions to be deleted and recreated without
affecting access to te function from the API gateway.


* Add concept of role to links

Links now have a "Direction" and a "Role"

The direction is either "inbound" or "outbound" at present.

"outbound" is the conventional use meaning an identity based security model when A is
given permission to access B.
"inbound" is a resource view where B permits A to perform operations -
resource permissions basically.

Roles allow the intent of the link to be defined. Permissions are
associated with the role. This way, the solution not only captures the
association of two components, but the nature of the association.

The mechanism is only configured for the API gateway right now. Separate
tickets will expand this for other uses.

This commit also includes fixes to ensure ECS task definitions and
externl links work correctly with the new occurrence structure.