Component configuration enhancements
Add ability to check configuration for children that are actually a
collection of subobjects. Subobjects can have default values defined on
higher level objects, with the resulting set of subobjects being a merge
of the objects available at each level. This greatly reduces the amount
of configuration needed for things like links and containers.

Add ability to add defualt configuration at the all of subcomponent level.

Update ECS and Lambda support to use the new facilities to completely
check container/function configuration.