Reduce use of global variables
With the move to use occurrences more heavily, the global omponent level
variables can be phased out.

This is a first cut at this - componentId and componentType are still
used in quite a number of places but other variables can be removed
in preparation for a more general move to use of occurrences.