Control paths in URLs for API Gateway (#204)
Include the version path in URL attributes - selection of version can
then be controlled via the links in the solution.

Provide an ability to turn this off for backwards compatability.

Also provide an attribute with the path that the implementation (e.g.
lambda) will see. The suspicion is that this varies depending on the
configuraiotn of custom domains. As the gateway knows its configuration,
this seems the best default place to control this information.