Correct detection of unresolvable links (#221)
Change behaviour of getLinkTarget to return an empty object if a link
isn't matched, rather than a partially populated one.

Add code to detect situations where getLinkTarget may return an empty
object. An exception is generated in this case, so the aim is to have
the template complete successfully so the Exception detection logic can
then report the failure to match the link.