Add control over signing fdqn
Support addition of a base path mapping when using cloudfront in front
of the api gateway so the fqdn used in the signature doesn't vary when
the api gateway is recreated.

Also remove the need for tier/component when generating a host name, as
these are already present in the occurrence structure.