Add ability to expand CIDRs
WAF is limited in the CIDR mask values it accepts so previously we have
explicitly enumerated IPs in the IP Address Groups.

However this can easily blow limits on security groups.

So for WAF, cidrs are expanded to meet its limitations while the better
support for variable masks in security groups means IP Address Groups
can be smaller.