ALB listeners are now subcomponents of the ALB
This allows each listener to carry its own certificate info and be
linked to via ECS containers.

It required the migration of the solution structure representing the
ports so a per-occurrence generation migration step has been added for
cases where older format solutions need to be migrated to support
backwards compatability. This only changes the internal structure used -
the assumption is the actual solution file can be edited separately to
align with the current spec.