Copy asFiles in prologue script
With settings now including entries for asFiles, the epilogue script can
now copy the specific files required, rather than having this logic in
manageStack. The logic is still there but will be removed once the move
to using settings is complete.

This commit also includes a few bug files and tiydyups.