Add more flexibility in managing app environment
Two patterns are now available.

1. Include the default environment and link settings
2. Explicitly specific the set of environment variables wanted from
those defined in the defualt environment.

The first is simplest but may result in unused variables in the
environment, depending on the setting overrides used.

The first is more work and requires a container fragment file but is
more precise. For spas, it is the only approach available as the
config.json is public and so some thought should go into what is
exposed.