Feature appsettings in occurrences (#260)
* Add generation of composite settings

Includes appsettings, credentials and asFile references

* Move solution configuration to a subdirectory

Configuration from the solution now under a "Solution" subdirectory.
This allows for other types of configuration under the top level
directory, e.g. settings.

* Move settings to occurrences and remove use of appSettingsObject

* Make exception string unique to COT

* Copy asFiles in prologue script

With settings now including entries for asFiles, the epilogue script can
now copy the specific files required, rather than having this logic in
manageStack. The logic is still there but will be removed once the move
to using settings is complete.

* Support application deployment subset templates

Now that all settings are available, it is possible to process all
application templates in one pass.

* Add default policies

Cover decryption and access to operations and data buckets. There is solution level controls as well.

* Support task/container level log groups

* Support disabling of component occurrences

* Manage build info as a separate group of settings

* Correct handling of CF Origin for API gateway

When no base path mapping is employed, use a reference to the api
gateway to work out the origin fqdn.

* Shared builds don't inherit asFiles

* asFile copying now in prologue script

Decrease the dependency of prologue and epilogue scripts on specific
functions within the gen3 script framework.

Instead, increase the amount of script generated by freemarker which can
make use of more generic support functions.

This also simplifies the steps when managing a stack by reducing the
default processing needed.

* Add more flexibility in managing app environment

Two patterns are now available.

1. Include the default environment and link settings
2. Explicitly specific the set of environment variables wanted from
those defined in the defualt environment.

The first is simplest but may result in unused variables in the
environment, depending on the setting overrides used.

The first is more work and requires a container fragment file but is
more precise. For spas, it is the only approach available as the
config.json is public and so some thought should go into what is
exposed.

* Reorder environment generation

While there's no guarantees, add link variables as the last step so
generated order matches existing templates to make comparisons easier.

It means link attributes will override and other environment settings of
the same name, so something to be aware of.