Remove usage functionality from IP Address Groups
Experience has shown that the usage capability in IP Address Groups can
be replaced by using more granular address groups and configuring lists
in the solution. This is also more explicit as it is obvious from the
solution what is actually happening, rather than indirectly based on
usage.

As part of this change, mandatory subchildren are now supported. they
will be ignored if the parent doesn;'t exist, but checked for where
there are candidate values to consider.