Cleanup temporary file management (#302)
All temporary files are now created in a temporary directory, and more
use is made of nested temporary directories to assist in debugging the
flow of function calls.

The composite_* files have been moved into a "cache" directory at the
root of the tree in preparation for improvements in the use of cached
information.

The templates for temporary files now used four X's. With only three,
there was occassionally a temporary file called "aux" created which is
not a valid file/directory name on windows.