asFile references are now relative to ROOT_DIR
To make it easier to have files coming from various directories once we
move to shared directories, change the processing of asFiles to be
relative to the root of the cmdb.

This allows the absolute path to vary (e.g. between workspaces for
different Jenkins jobs) but a prepared template should still work.

The commit also removes an unused function to assemble credentials,
which is done as part of assembling the settings.