Support environment and segment in names
Move to using the environment by default the segment only if it differs
from the environment and does not have the value of "default".

The next stage will be to change the cmdb layout to explicitly include
the environment in the tree to allow for multi-segment environments.

NOTE: This change may need adjustment to segment.json files where the
environment is not the same as the segment.