Introduce separate environment directories (#309)
Introduce an environment directory layer between product and segment
to support mult-segment products.

Rename appsettings to settings, solutions to solutionsv2, and
credentials to operations.

Support a reserved environment of "shared" to support sharing
across all instances of the same segment across environments, as well as
across all segments of all environments.

A segment of "default" will not appear in names, in much the same way as
an instance or version of "default" does not appear in names.

Support for appsettings and credentials has been removed as these have
been superceded by the settings support.

Refactor upgrade support to separate the upgrade process from the cleanup
process.

Remove the automatic migration of ALB config to ensure it is explicit
what is expected. Rework the LB support for EC2 and its config along the
same lines.

Use the ignoring of "dot" files to make ssh credentials invisible by
default to the app.

The commit also includes some general tidyup such as the use of FUNCNAME
as part of temporary directory creation, and removal of the functions to
manipulate CMDB files (this is now done in the affected templates).