Support option of creating lambda log group
While the log group name for lambda is currently fixed, the log group is
only created when the first lambda message is written.

This prevents the creation of log filters until the lambda runs.

To mitigate this, support the option of manually creating the group with
the AWS required name before creating the function.

Hopefully at some stage AWS will permit customer control of the lambda
log name, so this code positions the framework for this in the future.