Correct logic of linking to LB
In order that LB linking can be handled like any other subcomponent,
switch from linking on the basis of source port to linking on the basis
of the subcomponent id, which for an LB is the port mapping.

Also update link support to prefer the use of "PortMapping" to "Port"
for linking. This has meant adding the ability for checking multiple
Link attribute names when linking to a subcomponent to keep backwards
compatability with existing configs that are using "Port" rather than
"PortMapping".