Introduce reserved IPAddressGroups (#351)
Reserve all IP address groups starting with "_". Initially use these for
an open CIDR (_global) and the current segment nat ips (_segment).

Support a few variants with one or two "_" on either side as well.

Refactor the elastic search support accordingly.

We will likely add an _environment in future if all segments in an
environment need access to something.