ECS awsvpc networking and Network Load balancer Support (#359)
Add support for the awsvpc ECS container network mode

This creates an eni for each container in a service which bypasses the host network adapter. This removes issues with exposing the same port on multiple containers since each container has its own IP

Configuration - ECS Service Component NetworkMode - setting the network mode to awsvpc enables awsvpc for all containers within the service (defaults to "" which is the standard ecs network mode)

Add Target Type to LB

When using awsvpc mode you can't use the instance target type for an ALB or NLB target group since this just goes off the first adapter on the instance that responds. Instead you need to specify the IP address of the ENI for the specific container

Configuration LB Port SubComponent TargetType - set to ip to change the target group for a given port mapping to IP mode. Defaults to instance

Add Network Load balancer Security group configuration

When creating a network load balancer you cannot specify a security group on the load balancer itself. Instead it uses the security groups of its targets ( https://docs.aws.amazon.com/elasticloadbalancing/latest/network/target-group-register-targets.html#target-security-groups )

To ensure we are honouring the LB IPaddress groups we need to get the IPAddressGroups specified on the load balancer and apply them to the security group for the load balancing target. We also need to include the local VPC network so that the load balancer health probes can check the targets

Change container ingress rule from 0.0.0.0/0 to the security group of the load balancer.

0.0.0.0/0 works ok since its inside the VPC but since its easy for us to get the security group lets follow security best practice and only allow access to the container from the load balancer

Add _localnet reserved IPAddress Groups - returns the vpc local network

Add httpredirect standard port

The redirector container should only be treated as successful if its health check returns a 301 error not a 200, so we need to have a separate port/health check where this is defined.