ECS Remove Delegated Domain and add Task run Configuration (#362)
Delegated deployment was added to codeontap so that we could use the jenkins ecs agent plugin which ran ECS tasks on demand. This essentially required us to build an ECS tasks definition and pass it to the Jenkins server so it could create the task definition itself.

The plugin has been now updated so that you can specify an existing task definition to use and Jenkins will then use it.

This PR removes the DelegatedDeploy Task configuration and instead adds an overall capability for components to run/start an ECS task. Linking to a task with the role run will provide 2 Attributes

ECSHOST - The ARN of the ECS host the Task belongs to
DEFINITION - The family name of the TaskDefinition

DEFINITION is only available if the task has a fixed family name as this allows for the task to be updated without the calling service having to deal with a changing task definition. 

So a task now has a new configuration option

FixedName (booleen - default false) - This sets the family name to the occurrence name for the Task Definition. 

When you apply this option when Cloudformation updates the task definition instead of replacing the task definition it creates a new revision of the task definition instead. This is useful when you have reasonably fixed task definitions that don't change often.

The link also provided permissions to the component to Run and start the task on the cluster and also to stop any tasks on the cluster ( can't filter tasks based on their definition)