Add environment file support for lambda
Support an optional switch to place the environment settings in a config
file rather than as environment variables.

This avoids the limit on environment size in lambda.

Regardless of the switch, the core settings are still defined in order
that the bucket and settings prefix are available.

The variable CONFIG_FILE is also set with the relative filename of the
config file relative to the SETTINGS_PREFIX path.