Use fragment in preference to container
Expand the idea of a container fragment to apply to all components.

So the idea of a "fragment" is a piece of template (typically just macro
invocations) that can be associated with any component to customise its
configuration, typically to add security information.

As some container fragments may explicitly use the containerId variable,
leave it defined for now.

A cmdb upgrade process also needed to be added to rename any existing
container files to files.