Allow treating of runId as significant
Include the runId as metadata in every template

Treate runId differences as not significant unless specifically
requested.

Update "config as file" support to include runId in the S3 object to
trigger code updates when the config file contents change (assuming the
request asks for them to be treated as significant).

Align lambda code with the config as file changes.

NOTE: the api-gateway template already uses the runId to trigger redeploys
so will automatically be affected by the same switch flagging runId
differences as significant.