Refactor swagger definiton generation
Create a specific ftl file for the process of extending a swagger
definition file. Call the function it provides in the existing swagger
genertion process so it continues.

Over time, we will need to remove the Accounts and Regions sections of
the apigw.json file, as this is to be part of the dpeloyment process.