Dataset build Environment and IAM policy (#468)
Add a couple of extra dataset functions

Build environment configuration - Since data is only uploaded to the first environment where code is automatically built we need to allow this environment to be defined in the data set configuration. If the configured build environment matches the current environment the link lookup for the datasource is performed. If it doesn't only the registry details are provided.

IAM Policy - adds consume and produce roles for the data set. 2 roles are available the produce and consume roles. The produce role is only available in the build environment.

Not sure if this should be component configuration or an environment level configuration parameter.