Move swagger extension management to deployment (#466)
* Move swagger extension to deployment

Swagger extensions need various pieces of specific data injected into
them. For AWs SNS integration, we need a role arn so the existing
process won't work.

Shift processing of extensions into the template generation process.
This has the side benefit that the apigw.json file is now part of the
cmdb rather than the code repo, can be made environment specific, and
indeed can now use overrides and qualifiers.

The build process remains since it does validation, yaml to json conversion
and strips out any swagger language constructs that AWS doesn't support.

In order t get this working, introduce a pregeneration pass that can
generate a script which is run before any other template generqtion
happens. this script is used by the API gateway to retrieve the current
swagger specification and provide it as a base definition to the config
pass. Each definition is placed under the core.Id of its component in
case more than one component are included into a deployment unit.

* Simplify swagger definition file management

Move the bulk of the logic from the API Gateway template to a function
in createTemplate.sh - means a lot less escaping of quotes and brackets.

* Refactor swagger definiton generation

Create a specific ftl file for the process of extending a swagger
definition file. Call the function it provides in the existing swagger
genertion process so it continues.

Over time, we will need to remove the Accounts and Regions sections of
the apigw.json file, as this is to be part of the dpeloyment process.

* Source integrations from settings

* Integrate swagger generation with link processing

* Add role generation for swagger integrations

* Support generating a merged swagger spec

* Add composite definitions

Definitions need to be available to multiple subsets so must be included
in the cache.

* Further tweaks to swagger generation

Name the definition files after the component so multiple units can
update the same file.

Tweak the special iam unit processing to accommodate the need for
pregeneration before IAM to ensure roles for swagger integration are
correctly generated.

* Minor comment correction

* Correction to variable scoping