Add Support for User defined bootstrap scripts (#488)
Adds support for user defined bootstrap scripts. These can be used to perform additional hardening or configuration tasks on ec2/IaaS based components.

It also introduces the idea of a scriptstore, which can be used to add extra locations for scripts that could be used by different components, currently it just applies to user boot strapping but could be used for infrastructure level lambda functions and containers.