LB Fixes and Improvements (#490)
Features

Host Filtering - Adds a condition on a listener rule for the generated DNS hostname of the site. This prevents requests to the load balancer IP addresses which can be used for malicious probing. The option is applied on each port mapping and can only be used on ports with certificates configured

Changes

For Application Load balancers all listener rules are now created via CLI instead of cloudformation. This allows the rules to be managed easier and stops overlaps between cli and cloudformation which prevents updates

As part of this process all listener rules are removed before adding the new rules. This ensures that rules are cleaned up and only the configured rules will be applied. This will cause a brief (seconds) outage between the removal of the rules and the new ones being added
A default target group is created for all listeners and the default listener is always configured to use this rule

When a forwarding action is required for an application load balancer a separate target group is created along with a listener rule. This allows for conditions to be applied to all forwarding rules
Network load balancers can only use a single target group per listener at the moment so applications will receive the default target group arn when linking to a network LB

Targetgroup ARN look ups by linking components will receive the ARN via Attributes instead of a getReference lookup

Fixes

Align Security group name to the Id configuration which uses the parent details to create the property. This stops the security group from being recreated when a portmapping is renamed
Move all listener rule creation to cli based config, this stops issues when a cli listener rule has been created and a cloudformation based run is added with the same priority

Use the parent configuration details + priority for the listener rule Ids - if we move back to cloudformation this is required to ensure that rules are updated instead of trying to create a new rule with the same priority during updates

Add Health check configuration to all ports in master data

Update status code type for fixed response

Cleanup extended link configuration that is no longer required for latest lb config

Make sure listener rule and target group config is only applied on application load balancers

Make sure cli commands error when they break

Cleanup listener rules as part of deletion process