Add control over the ording of host name parts
A certificate or certificate behaviour object can now have a HostParts
array, listing the parts to be considered for a host name.

This list is used to drive the candidate parts considered, so anything
not in this list will not be included.

In addition, the booleans in IncludeInHost continue to apply for those
parts in the HostParts array. However, if the boolean is not present, is
it considered to have a value of true.

It is thus possible to override the host structure either via the
HostParts array or via controlling the flags for each part. By default,
the flags are used to not provide the Environment in the host name in
production environments.