Add support for usage plans
Allow the creation of a usage plan that can then be linked to from
users.

Create an API key for the user to be included into any usage plans to
which they are linked.

Move creation of some API Gateway resources to macros as they now occur
in multiple places.

A user is given permission to invoke any api to which the usage plan is
linked.