Introduce framework for typed links
A link can now have a "Type" attribute. It is only considered where the
link is external.

If present, it is used to determine what the target component type is
and generate the occurrence structure on this basis. this permits IAM
roles etc to still be generated even though the target is external.

Support needs to be explicitly added to any component state routine for
it to be supported. At present only SQS queues are implemented.