Component reference and id properties (#527)
Adds support for generate a markdown based reference document based on the component configuration defined in the *id.ftl files.

Adds a properties section to the component configuration objects which adds documentation or immutable attributes for a component

Properties have the following types

Description - A description of the component and what it does

Note - A informational note. Severity can also be set on a Note property to highlight the note appropriately

Providers - The providers that support the component

ComponentLevel - The level ( segment, solution, application, etc ) that the component is deployed with

Updated all Attribute name properties to Names instead of Name. Single string is still accepted and will be converted to a single item list during processing

To generate a reference you just need the GENERATION_DIR no cmdb is required

${GENERATION_DIR}/createReference.sh -t component
The file will then be available in ${GENERATION_BASE_DIR}/dist/reference/component-reference.md

An example of reference is available here
https://codeontap.readthedocs.io/en/latest/reference/component-reference/

More documentation needs to be added but this provides the foundations