Add active ink target detection
By default, getLinkTargets will only return active links - the target is
enabled and deployed. It will still return external links regardless.

This means there is no need to check for deployment in most situations.

Occasionally, such as with users, you may wish to have different logic
depending on whether the target is deployed or not, in which case the
Octivenly flag can be set to false, and isTargetActive() used to control
the logic flow.