Define stage variable for lambda even if its not deployed
In order to minimise the need for multiple deployment passes, define and
staging variables that hold lambda function names irrespective of
whether the lambda is deployed. Normally lambdas are sequenced after the
corresponding api to ensure the resource permissions are created
correctly, but currently that then means a second pass of the api
gateway template is needed to define the corresponding stage variable.