Repeat 1.3.1 upgrade (#599)
A bug in the original 1.3.1 logic meant that it did not rename all
files, particularly template files.

The logic of template file detection meant that if stack processing was
then performed, it was possible to end up with two instances of the
stack file in the cmdb.

The 1.3.1 logic has been upgraded to detect and where possible delete
the duplicate files, or flag the need for manual intervention where
differences exist.

In order to pick up cmds that were upgraded with the original 1.3.1
logic, 1.3.2 simply reruns the current 1.3.1 logic. In most cases, it
will be a no-op.