<refactor> GEN3 directory tree to separate state
Add support for semver range checking based on the syntax used by
node-semver.

Add the ability to constrain a cmdb upgrade to a particular version of
the GEN3 framework, mainly to allow a cmdb upgrade to be added to the
framework but not applied until the corresponding GEN3 upgrade is
active.

Add the upgrade to v2.0.0 of the cmdb, which separates state from
config/infrastructure. It has not yet been enabled to ensure the
refactoring still works with existing cmdb layouts.

Add functions to support switch to use of branches for releases.

Functions for context tree access now support the pre and post v2.0.0
cmdb directory formats. A few extra globals were added for locating
builds and state to assist with the transition between directory formats.

Note that once the v2.0.0 cmdb update is applied, build jobs will need
to be changed to reflect the new location, so they should use
SEGMENT_BUILDS_DIR when determining where in the repo to save build
references. There is an equivalent issue for gen3-automation that should
be applied with the issue to support the change of directory tree for
builds.

The v1.1.0 cleanup did not correctly clean up accounts repos so a v1.1.1
cleanup has been added that can handle multiple config/infrastructure
pairs in a single repo. This shold only affect accounts repos.

Add extra functions from gen3-automation/utility.sh so the two align.