Enable cmdb cleanup (#715)
Enable cmdb cleanup to be performed in addition to cmdb upgrade.

The cleanup code has been there for a while but was not invoked.

Cleanups are normally a ways behind their corresponding upgrades to give
time for the upgrade to percolate to all active projects. It is also
possible that cleanups will have been manually performed, in which case
the cleanup is a no-op.