<refactor> Force removal of "containerId" use (#728)
As part of switching to the term "fragment" in preference to the term
"container", remove the availability of the "containerId" variable from
within fragment files.

The switch of filenames has been in place for a while, and use of the
older variable should have been switched to fragmentId. This change will
identify any lingering use of the older term and thus may require
remediation of fragment files.