<bugfix> Align ids for cf access id and ops bucket"
Ensure the cloudfront access id pseudostack follows the same naming as
expected by the ops bucket. The ops bucket name was changedas part of
the transition to use of the baseline unit.

The code in baseline will now work for the old and new bucket id formats
provided the :wq