<refactor> Reduce rules needed for basic OWASP
WAf charges per ACL and per rule so good to minimise the number of rules
as far as possible. The basic OWASP checks can be combined in a single
rule since if any of them matches, the rule fires and the request can be
blocked.

Also rework the naming of rules and conditions to provide more informative
values in the console.