Remove depndencies on global tier/component
In readiness for dynamkic loading and simpification of the core
processing loop, ensure all templates only depend on the occurrence
variable and not directly on the tier ad component object.

Introduce some helper routines based on occurrence rather than
tier/component to simplify the cleanup process.

Ideally we want no references to tier or component directly, though
there are a few more things to address before that is the case.